move: type -> types
